# pkg/client/api/openapi.yaml
openapi: 3.0.2
info:
  contact:
    url: https://github.com/moov-io/customers
  description: |
    Customers focuses on solving authentic identification of humans who are legally able to hold and transfer currency within the US. Primarily this project solves [Know Your Customer](https://en.wikipedia.org/wiki/Know_your_customer) (KYC), [Customer Identification Program](https://en.wikipedia.org/wiki/Customer_Identification_Program) (CIP), [Office of Foreign Asset Control](https://www.treasury.gov/about/organizational-structure/offices/Pages/Office-of-Foreign-Assets-Control.aspx) (OFAC) checks and verification workflows to comply with US federal law and ensure authentic transfers. Customers has an objective to be a service for detailed due diligence on individuals and companies for Financial Institutions and services in a modernized and extensible way.

    Customer phone numbers and addresses are stored and partially used in KYC/OFAC validation. Arbitrary key/value pairs can be stored for a Customer. Documents and Disclaimers, and their acknowledgement are also stored under a Customer as they're accepted. Bank Accounts, which can be validated with micro-deposits currently, are stored under each Customer.

    ![](https://raw.githubusercontent.com/adamdecaf/customers/create-accounts/docs/images/customer.png)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Customers API
  version: v1
servers:
- description: Local development
...
paths:
...
  /configuration/customers:
    get:
      description: Retrieve current configuration for the provided namespace.
      operationId: getConfiguration
      parameters:
      - description: Value used to separate and identify models
        example: org342
        explode: false
        in: header
        name: X-Organization
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceConfiguration'
          description: Configuration was successfully updated
      summary: Get Configuration
      tags:
      - Configuration
    put:
      description: Update the configuration for the provided namespace.
      operationId: updateConfiguration
      parameters:
      - description: Value used to separate and identify models
        example: org342
        explode: false
        in: header
        name: X-Organization
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamespaceConfiguration'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceConfiguration'
          description: Configuration was successfully updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Configuration was not updated, see error(s)
      summary: Update Configuration
      tags:
      - Configuration
...
components:
  schemas:
    NamespaceConfiguration:
      example:
        primaryAccount: f6eddffd
        legalEntity: 5c20933f
      properties:
        legalEntity:
          description: customerID of legal entity for this Namespace
          example: 5c20933f
          type: string
        primaryAccount:
          description: accountID of funding/source account for this Namespace
          example: f6eddffd
          type: string
      required:
      - legalEntity
      - primaryAccount
...
    Error:
      properties:
        error:
          description: An error message describing the problem intended for humans.
          example: Validation error(s) present.
          type: string
      required:
      - error